update:: support height Input Multiline text
refactor ACRMultilineTextView
