+++ source/macos/AdaptiveCards/AdaptiveCards/Views/ACRMultilineTextView.xib
@@ -1,14 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.Cocoa.XIB" version="3.0" toolsVersion="17701" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" customObjectInstantitationMethod="direct">
+<document type="com.apple.InterfaceBuilder3.Cocoa.XIB" version="3.0" toolsVersion="20037" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" customObjectInstantitationMethod="direct">
     <dependencies>
         <deployment identifier="macosx"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="17701"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="20037"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <customObject id="-2" userLabel="File's Owner" customClass="ACRMultilineInputTextView" customModule="AdaptiveCards" customModuleProvider="target">
+        <customObject id="-2" userLabel="File's Owner" customClass="ACRMultilineTextView" customModule="AdaptiveCards" customModuleProvider="target">
             <connections>
-                <outlet property="contentView" destination="c22-O7-iKe" id="tYb-ha-9Wl"/>
                 <outlet property="scrollView" destination="6Xv-T3-ccx" id="71d-Ek-WrF"/>
                 <outlet property="textView" destination="Uet-pE-H0F" id="LGM-xC-zRf"/>
             </connections>
@@ -22,15 +21,15 @@
                 <scrollView borderType="none" horizontalLineScroll="10" horizontalPageScroll="10" verticalLineScroll="10" verticalPageScroll="10" hasHorizontalScroller="NO" translatesAutoresizingMaskIntoConstraints="NO" id="6Xv-T3-ccx" customClass="ACRScrollView" customModule="AdaptiveCards">
                     <rect key="frame" x="0.0" y="0.0" width="480" height="272"/>
                     <clipView key="contentView" drawsBackground="NO" copiesOnScroll="NO" id="hAP-XJ-GuW">
-                        <rect key="frame" x="0.0" y="0.0" width="465" height="272"/>
+                        <rect key="frame" x="0.0" y="0.0" width="480" height="272"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <textView importsGraphics="NO" richText="NO" verticallyResizable="YES" spellingCorrection="YES" smartInsertDelete="YES" id="Uet-pE-H0F" customClass="ACRTextView" customModule="AdaptiveCards">
-                                <rect key="frame" x="0.0" y="0.0" width="465" height="272"/>
+                                <rect key="frame" x="0.0" y="0.0" width="480" height="272"/>
                                 <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                                 <color key="textColor" name="textColor" catalog="System" colorSpace="catalog"/>
                                 <color key="backgroundColor" name="textBackgroundColor" catalog="System" colorSpace="catalog"/>
-                                <size key="minSize" width="465" height="272"/>
+                                <size key="minSize" width="480" height="272"/>
                                 <size key="maxSize" width="480" height="10000000"/>
                                 <color key="insertionPointColor" name="textColor" catalog="System" colorSpace="catalog"/>
                             </textView>
@@ -41,7 +40,7 @@
                         <autoresizingMask key="autoresizingMask"/>
                     </scroller>
                     <scroller key="verticalScroller" wantsLayer="YES" verticalHuggingPriority="750" horizontal="NO" id="TRj-dV-UY1">
-                        <rect key="frame" x="465" y="0.0" width="15" height="272"/>
+                        <rect key="frame" x="464" y="0.0" width="16" height="272"/>
                         <autoresizingMask key="autoresizingMask"/>
                     </scroller>
                 </scrollView>
